+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 4.1.0
+* Add `payment_reader_card_details` parameter to `Transaction.sale`
+* Add webhook sample for `GrantedPaymentMethodRevoked`
+* Add `skip_advanced_fraud_checking` to:
+  * `Customer#create` and `Customer#update`
+  * `PaymentMethod#create` and `PaymentMethod#update`
+  * `CreditCard#create` and `CreditCard#update`
+* Include `rexml` in Gemspec (fixes #214)
+* Fix incorrect error code constants (fixes #213)
+  * `PaymentMethodNonceConsumed` is now `93107`
+  * `CannotForwardPaymentMethodType` is now `93106`
+
 ## 4.0.0
 * Breaking Changes
   * Update gemspec to require ruby >= 2.6.0